[C++ SDK] Separated proto accessor for public and private SDKs
