+++ ydb/public/lib/ydb_cli/commands/ydb_service_scheme.h
@@ -6,10 +6,10 @@
 #include <ydb/public/lib/ydb_cli/common/format.h>
 #include <ydb/public/lib/ydb_cli/common/print_utils.h>
 #include <ydb/public/lib/ydb_cli/common/recursive_remove.h>
+#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/draft/accessor.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/draft/ydb_replication.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/draft/ydb_view.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/coordination/coordination.h>
-#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/proto/accessor.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/scheme/scheme.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/table/table.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/client.h>
@@ -69,7 +69,7 @@ static int PrintDescription(TCommand* self, EDataFormat format, const TValue& va
                  << "Use \"--format proto-json-base64\" option instead." << Endl;
             [[fallthrough]];
         case EDataFormat::ProtoJsonBase64:
-            return PrintProtoJsonBase64(TProtoAccessor::GetProto(value));
+            return PrintProtoJsonBase64(NDraft::TProtoAccessor::GetProto(value));
         default:
             throw TMisuseException() << "This command doesn't support " << format << " output format";
     }
