--- ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/proto/private.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <ydb/public/api/protos/draft/ydb_replication.pb.h>
-#include <ydb/public/api/protos/draft/ydb_view.pb.h>
-
-#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/draft/ydb_replication.h>
-#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/draft/ydb_view.h>
