+++ ydb/public/sdk/cpp/src/client/proto/ya.make
@@ -4,13 +4,8 @@ SRCS(
     accessor.cpp
 )
 
-CFLAGS(
-    GLOBAL -DYDB_SDK_INTERNAL_CLIENTS
-)
-
 PEERDIR(
     ydb/public/api/grpc
-    ydb/public/api/grpc/draft
     ydb/public/api/protos
     ydb/public/sdk/cpp/src/client/params
     ydb/public/sdk/cpp/src/client/value
